requirements.txt
pika
SQLAlchemy
psycopg2-binary
python-dotenv
